Create test activity for adding a 2nd SurfaceView
This activity is initialized with a FrameLayout containing a SurfaceView
on the bottom layer and an opaque white view on top of that. A button
allows the user to programatically add a second SurfaceView on top. The
Surface associated with the bottom SurfaceView is filled with red.
Notice the red flash that sometimes occurs on addition and removal of
the second SurfaceView.